applets/clock: Only update the label when the content would change
It is fairly cheap for us to compare the two times as they would
appear when rendered to the panel. However, changing the label will
then force a resize regardless of whether we redraw, which is causing
the top level to undergo full resize events every second, which is
costly.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>