panel: Add support for Left + Right panels
This is still primitive and early but it does basically work. We'll still
need to resolve the overlaps between the top/bottom and side panels by
removing priority from the side panels but this is a simple enough task.

Additionally we'll need to make Raven pop out from underneath a right panel
if it exists.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>