plugin: Add a closed signal to the BudgiePopover
In the interests of a simple port from GtkPopover to BudgiePopover, we
provide a new `closed` signal, which is emitted after the popover has
been unmapped.

This is an explicit signal of BudgiePopover, so that consumers of the
API do not need to worry about the internals of GtkWindows and their
associated signals/lifetimes.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>