applets/clock: Adapt to vertical panels, stop using "big"
The big font just looks awkward as hell. However the "normal" font is still
going to be too big for the vertical panel, so in that instance we'll then
use "small" markup.

For final brownie points we give the panel applet a new style class to allow
CSS themers to go further.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>