panel: Fix clipping when using 34/37 icon sizes in icon tasklist
Without this we'll see clipping when we choose a 32px icon as we have
some internal padding within the buttons, so we must respect a minimum
content margin of 5px in relation to the icon sizes. This is most noticeable
on the bottom panel where a resize goes beyond the screen edge.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>