panel: Start paving the way for a dock mode
In essence, when in dock mode we only want to use the minimum amount of
space within the window. For a horizontal window we align the body to
the start of the window, and for verticals we vertically align to the
top.

The next steps will involve implementing a new positioning system to
dynamically update the window position based on the available geometry.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>