plugin: Skip invisible widgets
When iterating the associated parent -> popover table, we must check
that the widget of interest is actually visible. For example, the bluetooth
indicator may hide itself if not supported, but we'd still check the geometry.
In most cases this is broken, so we ensure we do not count it as part of our
validation.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>