panel: Teach Budgie Panel about active window intersections
This will be used at a later stage to implement an automatic hiding system,
whereby a nearby window will set the panel into an intersected mode, which
will usually mean we want to hide the panel.

The actual "hide" portion will need to wait, because we also want to be able
to peek the dock by moving our mouse into that region, which will likely
require the use of invisible struts and an enter-notify.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>