Implement initial autohide support for the Budgie Panel
This allows us to follow either an "automatic" or "intelligent" approach
to automatically hiding the panel. In "automatic" mode we'll simply hide
the panel during a screen occlusion event, i.e. something went full screen
or maximized or such.

In "intelligent" mode we check for both the simple screen occlusion, as well
as proximity of the windows to the panel. When a window is too close, the
panel will attempt to hide to dodge the window completely.

The next steps include implementing a "re peek" mode, i.e. when the mouse
enters the panel region to bring it back up for display temporarily.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>