panel: Modify input regions instead of hiding actual window
To ensure we can still access some kind of events system, we'll nuke most
of our X11 input shape so that events pass straight through our unrendered
window when hidden. When restored in autohide mode, we'll just get our old
input shape back.

Currently this doesn't yet allow us to get the "peek" event, as we'll need
to reserve a line along the furthest screen edge in the direction of the panel
to get the enter-notify-event. This will allow us to determine whether to
start a new show again. This stuff will be handled in subsequent commits,
along with basic throttling for the show/hide mechanism.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>