theme: Build CSS at build time, instead of in tree in git
Typically managing our theme has been a collosal pain due to having to
rebuild it in source and commit it separately, which leads to awful issues
for people trying to backport patches in git.

This change requires sassc as a build dependency, but ensures we only
build .css files at build time during packaging, making life easier for
everyone involved.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>