plugin: Prevent "grab wars" with GtkPopover
During some modal dialog experiments on the savin-raven branch, it
was determined that a grab war would ensue if one activates the Budgie
Menu while a modal dialog within our own process is active, i.e. the
modal dialog and the popover get into a never ending fight, leading
to a local denial of service.

This isn't part of any Budgie release and as such doesn't represent
any security risk, as this is unreleased, unstable code, on the main
development branch of Budgie.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>