panel: Fix random positioning of applets when moving them to the right
A very specific bug exists whereby moving applets from the left to the
right would upset the internal ordering of the panel. As an example, moving
the User Indicator all the way to the left, and back into the right. As
the indicator enters the right, the positions were thrown off for being
too high in the right most box, causing widgets past the index to become
unsorted.

Additionally after mild fixes it was found that with 2 applets in the
center, the left most applet when moved twice, i.e. once past the first
applet, and second into the right box, would jump past the first applet
in the right box.

This change fixes the behaviour by enforcing our positioning once the
signal noise has abated.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>