panel: Add confirmation dialogs for dangerous actions
When removing a panel or applet, we'll chuck up a simple confirmation
dialog and ask if the user really meant to do this. For power users this
dialog could quickly become annoying, so we also add a tickbox where the
user can request to not be nagged anymore by a confirmation prompt.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>