panel: Start working on a new LayoutManager
To make PanelManager less complicated, we'll have a dedicated LayoutManager
that knows how to read and write layouts, and from all the various locations
they need to be.

Eventually this will support two way serialisation and allow us to dump
current panel configurations, as well as reading and resetting to a named
panel confguration at will.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>