settings: Ensure we show the new settings applet
This fixes #1045 and other related issues by ensuring the new settings
portion of the applet is actually shown. i.e. when we switch between the
panels and their settings, the stack is able to switch to the new child.

Without this change, the stack will fail to update the currently visible
page and we'll end up mangling the _old_ settings for an unrelated applet.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>