daemon/tabswitcher: Adjust UI automatically and enforce proper layout
This ensures the tabswitcher grows as required, firstly setting the absolute
minimum size needed, and ensuring we allow *up to* 8 children per row which
will depend on the current set we're looking at.

This ensures that regardless of theme sizes we always have at most 8 elements
in a row, and we also allow the tabswitcher UI to properly center when there
are less elements to deal with, as opposed to having everything bunched up
at the start.

Lastly, we bump the icon size up to 48 to make use of the horizontal space
a bit better.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>