Restore ActionMenu functionality
Opening this for discussion.

This PR restores the WnckActionMenu functionality.
The menu is activated with Ctrl+RightClick.
When grouping is enabled and there is more than one window in a group, it
will only display the menu for the current active window from that group. (no active window in group = no menu)
It will always display the menu when there is only one window in the group.