applets/icon-tasklist: Rely on GtkImage due to icon scaling being bork
GtkImage relys on private GtkIconHelper and GtkScaling which aren't part
of the public API, so GtkIconTheme will continously fail for us even when
using the `_for_scale` APIs, as the GdkPixbufs will never scale correctly.
Instead, bite the bullet and extend GtkImage with custom rendering on top
of a cairo surface much like we do in Raven.

This fixes #1422

Signed-off-by: Ikey Doherty <ikey@solus-project.com>