Implement broader Raven DND support and tweak Clear Notifications.
Implements a dnd_enabled state in RavenIFace, as well as appropriate functionality for setting, fetching, and being updated on DND state changes. This is utilized by our Notifications / Raven Applet, which will now update the icon based on the DND state (defaults to not being in DND mode).

Additionally, tweaked the icon for enabling / disabling DND to being notification-alert-symbolic, to be consistent with the applet, and disabled being notification-disabled-symbolic.

This commit also tweaks the behavior of showing the Clear Notifications button, which will now only show if there are actual notifications to clear.

Added a .vscode workspace configuration to enforce spaces for Vala.