Implement keyboard shortcut for clearing notifications.
This commit implements Super+C functionality for clearing all notifications. To do this, we've implemented a ClearNotifications function in RavenIFace that sets the notification count to zero, marks them all as read, then proceeds to call our ClearAllNotifications signal.

This signal is utilized by a signal bindign (on_clear_all) in NotificationsView, which will clear all items from the listbox and update the child count.

This resolves #709.