Implement new Sound Output and Input Widgets in Raven.
This commit implements the brand new sound output and input widgets, where the Sound Output Widget provides per-app volume control, global control, and output device selection. The Sound Input Widget provides device control as well as microphone volume control.

When no apps are playing audio, we'll prompt the user to start listening to some music (checking for audio/vorbis support). When apps are playing, they appear in a list of AppSoundControl widgets, where we attempt to find the "best" or most relevant application icon from the icon theme, not necessarily the one Gvc may provide (in rare instances). We'll also attempt to filter out apps like System Sounds, WebRTC Engine (Discord), event-based volume change sounds, and more.

This commit additionally makes tweaks to the RavenExpander and its relevant components, to fix issues where it was always assumed that each widget would default to be expanded and thus not properly change its state.

Another feature in this commit is the MprisGui now setting a class for the player or its desktop entry (such as "spotify"), which should hopefully improve the level of customization for theme authors.

Updated all copyrights to 2015-2018.