Implement mute / unmute functionality in AppSoundControl
This commit introduces mute and unmute functionality in AppSoundControls, so you can perform said action on individual apps. This works by tracking the volume of the app and setting it to 0 or the volume prior to mute, since GVC set_is_muted and is_muted functionality doesn't work or reliably work. During construction, we'll check if the volume of the stream is equal to or below the minimum threshold, and set it to mute-by-default if it is.

Also note that this change introduces a new Gtk.Box as a "header" for each sound control, apologies in advance to theme developers for any changes necessary.

Lastly, this commit also works towards resolving some bugs around clearing all notifications. It appeared that during clearing, performing steals against the HashTable would break the foreach call and cause it to not continue to recursively clear notification groups. We now have a performing_clear_all check and perform a steal_all after it (in clear_all()). We'll still perform a steal and update_child_count if we're not performing a clear all during our function handler the dismissed_group NotificationGroup signal.