Implement support for showing all windows on IconButton click.
This isn't something I want to force as a default, as I'm sure there are plenty of usecases where we may not want to show all windows on click (non-incognito v.s. incognito browsers as an example), however I still want to provide this as an option for our users.

This effectively resolves #1409 by providing an option for an IconTasklist to have icon buttons activate all windows during click.