Fix crash on apps with no valid class group.
This commit should resolve #1630, which is a crash-related issue caused by some windows such as glxgears not having a Wnck.ClassGroup. We now type check the group and fallback to the app / window name.

This commit starts work on changes to better support (or at least attempt to disable some grouping) of Chrome + Chrome Apps. It isn't noticeable with this commit alone, as other changes are local / WIP, but I didn't want to unstage those chunks.