Ensure we still create an IconPopover even when we fail to get DesktopAppInfo.
During my testing with Firefox Developer Edition, I noticed that since it's just provided as a normal tarball and ran from a directory, it has no respective DesktopAppInfo (no desktop file). As a result, it'd fail to create the IconPopover. Now, we'll create it and only show the Close All button, since Favorite and Launch New Instance buttons require the AppInfo to exist.