Make some refinements to AppSoundControl.
This commit changes the behavior of the mute button to essentially act as a temporary toggle for the user to mute the volume while still interacting with the volume slider. This could be useful if at launch, the volume of an application is too loud, and the user wishes to quickly mute it, lower the volume, then unmute at a safer volume.

Additionally, if the volume is set to 0 on the closing of an application, we define this as being "pre-muted" and allow for the immediate raising of the volume with needing to unmute first.