Perform icon lookups for non-standard applications.
Ensure we perform a lookup of the Icon key/val, if it exists, of an application which has a respective DesktopAppInfo during our on_notification_closed func. This resolves #1669 where an application (GNOME Pomodoro) uses the standard file naming convention set force in the Desktop Entry specification, however does not do the same for their icon.