Implemented mechanism for hiding Sound widgets when there are no devices
Implemented a mechanism for dynamically showing and hiding Sound Output and Sound Input widgets based on if there are devices plugged in of that type, while still respecting the setting defined in Budgie Desktop Settings.

An example of this being used is when there are no input devices plugged into the machine. In this case, we should not show the Sound Input widget, regardless of the setting in Budgie Desktop Settings. However, if we plug in an input device, we should show the widget again if permitted by our setting. The same goes for device removal. If we no longer have a device of a specific type, we should hide it (output and input respectively).

This resolves #1685. Tested by unplugged my Logitech C920 as well as my Blue Yeti, plugging / unplugging the Yeti. Tested in cases of new session startup (new panel) and existing panel (dynamic).

This commit also implements a very basic RavenWidget class with a set_show function for toggling show / hide based on requested state. Drastically simplifies redundant show / hide widget calls.