Refine class-based window grouping and blacklisted window types.
This commit adds further refinements to our class-based window grouping logic, determining when we should add a window to a popover, ignore a window altogether, etc.

Firstly, we're now using AbominationRunningApp to track the proper group of the first window set in a class group provided to IconButton. Whenever a new window is opened, we then check if this class is LibreOffice and if so, we generate an AbominationRunningApp for the new window for group name comparison. This is useful since each LibreOffice window has its own Wnck.ClassGroup with different Ids but the same name. AbominationRunningApp is able to determine whether it's in fact libreoffice-writer v.s. libreoffice-impress, as an example, and ensure we more consistently group related LibreOffice windows together. Resolves #1682.

This behavior may be extended to Chrome and Chrome apps in the near future.

Secondly, this commit ensures we ignore not just DESKTOP type windows like Nautilus, but also DIALOG and SPLASHSCREEN type windows as well. DIALOG is useful for Impress and its template selection, since it'll show up under Writer otherwise. SPLASHSCREEN helps us to ignore GIMP and soffice splash screen.