Refined app name and icon logic for AppSoundControl.
This commit refines our previous method for determining the usable name and icon for an AppSoundControl. We now attempt to get the icon for an app before using the stream name, but still fallback to applications-multimedia if we fail to get any icon.

Additionally, this commit strips out any verbose "ALSA plug-in [blah]" text and just replaces it with "blah". A couple real world examples would be Audacity and mocp. Resolves #1692.

For Audacity, we not only now successfully get the desktop file successfully.

Any apps where we successfully get an icon (like Audacity, Kdenlive, etc.) we'll also now titalize. So previously you'd see kdenlive, now it's Kdenlive.