Implement Center Windows option in Budgie Desktop Settings -> Windows.
This commit adds support for enabling / disabling the capability to automatically center newly launched windows. Implemented a new schema option for this so it'll carry over for Budgie 11 and we're not permanently relying on Mutter's setting for it. However, we do respect the Mutter setting and apply that by default.