Rewrite of Caffeine Mode functionality (and most of its Applet).
This is a rewrite of various Caffeine Mode functionality that was primarily housed in the Caffeine Applet in a prior commit. In this commit, such functionality has been moved to the Budgie Daemon, where most of the other settings are handled. This is also in part to the fact that literally all the Caffeine Applet settings have been shifted to the window manager schema (though no WM code is actually manipulated here). In theory this means that third parties / downstreams can, if they wish, implement their own Caffeine Mode enable / disabling functionality elsewhere if they wish, while not being in conflict with the "core" Caffeine Applet.

The biggest reason for moving it to Budgie Daemon was our ability to then reliably have an Application (object) and have Notify reliably init. This was to enable us to resolve issue #1695 and is also accompanying a special case handler in NotificationGroup for treating the Budgie Daemon + Caffeine icons group name as just "Caffeine Mode".

Additionally, as part of this move, we're now using a memory-backed store for certain power and session Settings objects in the daemon. What this *should* mean alongside our disabling of Caffeine Mode at the start of the daemon, you shouldn't be accidentally messing up your power settings on a more permanent basis if Caffeine Mode isn't turned off before your session ends.

This commit also handles an error where we'd assume an Icon existed for an application just because the DesktopAppInfo was found during our Notify notification creation, when that may not necessarily be the case. Also documented *why* we do the settings_app_name variable setting so special.