Rewrite scroll functionality for IconButton.
The previous IconButton scroll functionality (including the get_previous_window / get_next_window) only checked the items in the current window's Wnck.ClassGroup, which is problematic for windows which do not share them, such as LibreOffice.

To resolve this, we're now checking our ids from the popover (since we know better than Wnck alone) to properly iterate across the windows. This resolves #1703.

Additionally, changed behavior so an UP direction scroll is next instead of previous.