Reset session and power values if Caffeine Mode is on at launch.
In the event of an improper shutdown, during the next session start we will reset the session and power settings if Caffeine Mode is still on. It's not ideal, however until we have a reliable way to have session-based and persistent key/values and use this interchangeably for settings, this is frankly the best way of doing it. Hopefully should address #1706.

This commit also does an immediate reset values during do_disable and do_disable_quietly, hopefully to ensure values are immediately reset prior to a shutdown. Lastly we also perform a do_disable_quietly on End Session dialog closure in addition to open.