Fix old IconButton logic for getting the current active window.
Previously, we would not attempt to get the target window if there wasn't a current window to do ID checks against. Now we will just default to 0 (first item in the current ids list) if we've determined there is no current active window. This can occur if all windows are minimized and you attempt to scroll to select a window. Resolves #1711.