Implement option to hide Built-in Theme item from Budgie Desktop Settings.
Realistically, the current built-in theme is neither maintained, nor recommended for usage. Since I do not want to suddenly shift people away from using the Built-in theme, if they are still choosing to use it, I've implemented a schema key called "show-builtin-theme-option". By default, this option is enabled, to preserve current behavior, however this option *may* be disabled in the Solus-specific branding package to discourage the use of the unsupported / unmaintained internal styling.

Among many other reasons, one reason this was implemented was wonky SoundWidget (output-type) input marks. To expand on ensuring these marks don't cause Raven to derp, we now have an update_input_draw_markers function which will check both if we have a supported GTK+ Theme (basically, not Arc*) or if we have Built-in Theme disabled. If that is the case, we show the marks (if audio overdrive is enabled), otherwise we keep them hidden.