Implement Gtk (Widget) and Icon Theme blacklist.
Until Budgie 11, where there will be a separate mechanic suggested for theme developers to use to expose support for Budgie, I'm implementing a blacklist of Gtk and Icon themes which are known to not work under, or cater towards, Budgie. This blacklist will ensure that users will not be able to choose themes that negatively impact their experience of the Budgie environment.

Furthermore, I have changed the volume slider mark on the Sound (Output) widget in Raven to have no text, instead of 100%. This means themes such as HighContrast will only show one mark instead of two, so it won't arbitrarily expand the volume slider. Theme developers are encouraged to take a look at Plata for what the suggested, best implementation for how the slider should look.