Revert changes for attach-modal-dialogs and edge-tiling setting.
The expectation for Budgie is to have a minimal reliance of the GNOME Stack leading up to Budgie 11, where a full breakaway will occur. To continue to enable this as well as ensure the greatest likelihood for settings to carry over from Budgie 10.5 to Budgie 10.5.1, as well as when we transition from Budgie 10.x to Budgie 11, we should continue to utility Budgie keys, even if they are duplicates of Mutter keys.

Where necessary, we will set the respective GNOME Desktop / Mutter keys.