Implement single-window tooltip info setting for IconTasklist -> IconButton.
In instances where we have a single window open for an IconButton, instead of showing the application or class name we will, if possible, set the current window title instead. This is done whenever you hover over the button to actually show the tooltip. For multi-window instances, we'll display the application name like we did prior to this commit.