Implement window invalidation in AbominationRunningApp
For instances where a window is closed but is still referenced by AbominationRunningApp.window, we should handle invalidating it and checking for additional windows in its class group that should replace it and be tracked.

To do this, I implemented an invalidate_window function that will check if the provided window has a matching xid and if so, we'll check for new windows that shouldn't be skipped in its class group. Once we find one, we'll set it as the new window, indicate we found a window, and immediately return.

If we find a window, we'll update our bindings, name, and icon. If we don't, we'll attempt to fall back to the DesktopAppInfo display name for it, and set that as the name (in effect it'll be immutable at that stage).

Also moved our bindings to a set_window function in AbominationRunningApp since there is now the possibility of changing the window in it. This is specifically designed to handle cases like Discord where they first launch a Discord Update window, which is then closed and replaced with a proper Discord window, that we'll want to track. We're also handling earlier returns for update_group and update_name if the window, whcih we return in those functions, is null.

In IconButton, we're utilizing the name_changed event in AbominationRunningApp to trigger set_tooltip and in the Wnck.Screen.window_closed signal hander we're checking if there is an AbominationRunningApp and if so, triggering its invalidate_window.