Drop support for pre-GNOME 3.34 Stacks
None of these are receiving adequate testing and patches from the community and it isn't realistic for us to both continue supporting old stack releases and ensuring we have a codebase that is maintainable and easy to iterate on until we release Budgie 11.

In addition to this, not providing gnome-bluetooth-1.0 and this being able to disable Bluetooth is not an officially supported configuration and thus this option has been removed.