Consolidate WindowType checking code into Abomination.
The WindowType checking across Abomination, IconButton, and IconPopover was getting out of hand. It grew organically to solve various issues over time and the time for it to be reigned in and consolidated into Abomination is now.

In addition to this reigning in, we've added UTILITY-type windows to the list of "disallowed" window types that shouldn't create IconButtons or show up in the IconPopover. SPLASHSCREEN-type windows should be blocked more consistently as well. Furthermore, we added is_skip_pager to existing usage of is_skip_tasklist.

May resolve #1822, awaiting validation.