Drop use of Queue and Cancellable.
Instead of using a Queue (which isn't thread safe anyways), just update a reference to a NotificationWindow since we only really cared about the head of the Queue to begin with.

Dropped Cancellable usage for async pixbuf scaling. There hasn't really been a scenario yet where we haven't been able to scale an icon within the amount of time before a NotificationWindow is removed. Open to adding it back if we run into problems.